+++ TeamProject/static/js/board/post/render.js
@@ -3,6 +3,8 @@ import * as FETCH from "./fetch.js"
 import * as EVENT from "./event.js"
 import * as MAIN from "../main.js"
 import * as COMMENT_EVENT from "../comment/event.js"
+import * as COMMENT_INDEX from "../comment/index.js"
+import * as COMMENT_RENDER from "../comment/render.js"
 
 export function input_post_window() {
     const html = '<div class="input__on"><input type="text" class="input__subject" maxlength="25" placeholder="글 제목을 입력해주세요" >' +
@@ -66,13 +68,10 @@ export function input_post_window() {
       '<div class="comment_last"><input type="button" class="btn_go_main" value="목록으로" /></div></div>';
     post_ele.innerHTML = html;
     post_img(post.post_img_filename);
-    //수정 삭제 그릴지 판단 : 현재로그인 한 user.id 와 post.id가 같은지 비교하고 같다면 수정삭제를 할 수있는 버튼을 볼 수 있게함
-    //수정
-    if (post.userid != userid) document.querySelector('.infoTop_buttons').style.cssText = ' display: none';
   
-    // EVENT.handle_update();
+    if (post.userid != userid) document.querySelector('.infoTop_buttons').style.cssText = ' display: none';
+
     EVENT.update_post();
-    // EVENT.handle_delete();
     EVENT.delete_post();
   }
 
@@ -86,59 +85,59 @@ export function input_post_window() {
     }
   }
 
-  export function post_comment_list(comment, user_data, login_currentUserData) {
-    let comment_html = `<div class = "comment_item" id="comment_id_${comment.id}"><div class="comment_top">` +
-      `<img src="${'http://127.0.0.1:5000/static/img/profile_img/'+user_data.profile_img}">` +
-      `<div class = "comment_info">` +
-      `<span class="comment_nickname">${user_data.nickname}</span>` +
-      `<div class="comment_buttons1">` +
-      `<input type="button" id = "comment_likes_${comment.id}" value="추천 ${comment.like_num}" />` +
-      `<input type="button" id = "comment_report_${comment.id}" value="신고" />` +
-      '</div>' +
-      `<span class="comment_date">${MAIN.calc_date(comment.create_date)}</span>` +
-      '</div>';
+  // export function post_comment_list(comment, user_data, login_currentUserData) {
+  //   let comment_html = `<div class = "comment_item" id="comment_id_${comment.id}"><div class="comment_top">` +
+  //     `<img src="${'http://127.0.0.1:5000/static/img/profile_img/'+user_data.profile_img}">` +
+  //     `<div class = "comment_info">` +
+  //     `<span class="comment_nickname">${user_data.nickname}</span>` +
+  //     `<div class="comment_buttons1">` +
+  //     `<input type="button" id = "comment_likes_${comment.id}" value="추천 ${comment.like_num}" />` +
+  //     `<input type="button" id = "comment_report_${comment.id}" value="신고" />` +
+  //     '</div>' +
+  //     `<span class="comment_date">${MAIN.calc_date(comment.create_date)}</span>` +
+  //     '</div>';
   
-    if (login_currentUserData.id == comment.userid) { //수정 삭제 그릴지 판단
-      comment_html = comment_html + `<div class="comment_buttons2">` +
-        `<input type="button" id = "updateComment__${comment.id}" value="수정" />` +
-        `<input type="button" id = "deleteComment__${comment.id}" value="삭제" />` +
-        `</div>`;
-    }
+  //   if (login_currentUserData.id == comment.userid) { //수정 삭제 그릴지 판단
+  //     comment_html = comment_html + `<div class="comment_buttons2">` +
+  //       `<input type="button" id = "updateComment__${comment.id}" value="수정" />` +
+  //       `<input type="button" id = "deleteComment__${comment.id}" value="삭제" />` +
+  //       `</div>`;
+  //   }
   
-    comment_html = comment_html + '</div>' +
-      `<p class="comment_content">${comment.content}</p><hr></div>`;
-    return comment_html;
-  }
+  //   comment_html = comment_html + '</div>' +
+  //     `<p class="comment_content">${comment.content}</p><hr></div>`;
+  //   return comment_html;
+  // }
 
-  export async function post_comment(comments) {
-    let text = '';
-    const login_currentUserData = await FETCH.fetch_userinfo();
+  // export async function post_comment(comments) {
+  //   let text = '';
+  //   const login_currentUserData = await FETCH.fetch_userinfo();
   
-    for (let i = comments.length - 1; i >= 0; i--) {
-      const user_data = await FETCH.fetch_getUserdata(comments[i].userid);
-      text += post_comment_list(comments[i], user_data, login_currentUserData);
-      //수정
-    }
+  //   for (let i = comments.length - 1; i >= 0; i--) {
+  //     const user_data = await FETCH.fetch_getUserdata(comments[i].userid);
+  //     text += post_comment_list(comments[i], user_data, login_currentUserData);
+  //     //수정
+  //   }
     
-    document.querySelector('.comment_list').innerHTML = text;
-    COMMENT_EVENT.add_comment_likes();
-    COMMENT_EVENT.add_comment_report();
-    if(INDEX.is_comment_exist(login_currentUserData.id,comments)){
-      COMMENT_EVENT.update_comment();
-      COMMENT_EVENT.delete_comment();
-    }
-    document.querySelector('.comment_num').innerText = `${comments.length}개의 댓글`;
-  }
+  //   document.querySelector('.comment_list').innerHTML = text;
+  //   COMMENT_EVENT.add_comment_likes();
+  //   COMMENT_EVENT.add_comment_report();
+  //   if(COMMENT_INDEX.is_comment_exist(login_currentUserData.id,comments)){
+  //     COMMENT_EVENT.update_comment();
+  //     COMMENT_EVENT.delete_comment();
+  //   }
+  //   document.querySelector('.comment_num').innerText = `${comments.length}개의 댓글`;
+  // }
 
-  export async function post_comment_update(id){
-    const ele = document.querySelector(`#comment_id_${id}`);
-    const ele_textarea = MAIN.get_htmlObject('textarea', [], [], ele.querySelector('p').innerText);
-    ele.replaceChild(ele_textarea, ele.childNodes[1]);
-    const button = ele.querySelector(`#updateComment__${id}`).parentNode;
-    const new_button = await MAIN.get_htmlObject('input',
-      ['type', 'id', 'value'], ['button',`updateComment__${id}`, '완료']);
-    button.replaceChild(new_button, button.childNodes[0]);
-  }
+  // export async function post_comment_update(id){
+  //   const ele = document.querySelector(`#comment_id_${id}`);
+  //   const ele_textarea = MAIN.get_htmlObject('textarea', [], [], ele.querySelector('p').innerText);
+  //   ele.replaceChild(ele_textarea, ele.childNodes[1]);
+  //   const button = ele.querySelector(`#updateComment__${id}`).parentNode;
+  //   const new_button = await MAIN.get_htmlObject('input',
+  //     ['type', 'id', 'value'], ['button',`updateComment__${id}`, '완료']);
+  //   button.replaceChild(new_button, button.childNodes[0]);
+  // }
 
 export async function post_update(post) {
     const user_data = await fetch_getUserdata(post.userid);
@@ -202,7 +201,6 @@ export async function post_update(post) {
           preview.appendChild(div); //이미지태그 그리기
         } else alert('이미지파일만 업로드가능합니다');
       }
-      // EVENT.handle_inputFileDelete();
       EVENT.delete_upload_file_in_post_input();
     }
   }
