+++ TeamProject/templates/post.html
@@ -82,18 +82,24 @@
 <script src="https://kit.fontawesome.com/8625eebbbc.js" crossorigin="anonymous"></script>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
 
+<!-- URL JS -->
 <script type="module" src="/static/js/config.js"></script>
+
+<!-- BOARD JS -->
 <script type="module" src="/static/js/board/event.js"></script>
 <script type="module" src="/static/js/board/main.js"></script>
 <script type="module" src="/static/js/board/render.js"></script>
 <script type="module" src="/static/js/board/fetch.js"></script>
 <script type="module" src="/static/js/board/router.js"></script>
+<!-- LIST JS -->
 <script type="module" src="/static/js/board/list/render.js"></script>
 <script type="module" src="/static/js/board/list/index.js"></script>
 <script type="module" src="/static/js/board/list/event.js"></script>
+<!-- ASIDE JS -->
 <script type="module" src="/static/js/board/aside/render.js"></script>
 <script type="module" src="/static/js/board/aside/main.js"></script>
 <script type="module" src="/static/js/board/aside/event.js"></script>
+<!-- AUTH JS -->
 <script type="module" src="../static/js/Auth/render.js"></script>
 <script type="module" src="../static/js/Auth/main.js"></script>
 <script type="module" src="../static/js/Auth/event.js"></script>
